Restructure files and folders
