+++ src/modules/product/product.routes.js
@@ -1,6 +1,6 @@
 const router = require('express').Router()
 const { createProduct, deleteSingleProduct, interestedUsers } = require('./product.controller')
-const {verifyUser} = require('./user.controller')
+const {verifyUser} = require('../user/user.controller')
 
 router.post('/create', verifyUser, createProduct)
       .get('/:productId/interested-users', verifyUser, interestedUsers)
