+++ src/modules/user/user.validator.js
@@ -1,7 +1,5 @@
 const Joi = require("joi")
 
-const { validator } = require("../utils/helpers")
-
 const signUpValidate = Joi.object({
     fullname: Joi.string().trim().required(),
     email: Joi.string().trim().required(),
